add telegram.go
